push from marc refonte des interfaces vendors

+++ lib/cubits/switch_page/switch_page_cubit.dart
@@ -5,7 +5,7 @@ import 'package:e_com_app/cubits/switch_page/switch_page_state.dart';
 
 class SwitchPageCubit extends Cubit<SwitchPageState> {
   SwitchPageCubit() : super(SwitchPageState.initial()); // Par défaut, on commence à la page 0 (Dashboard)
-  void switchPage(int pageIndex) {
+  void switchPage(double pageIndex) {
     emit(state.copyWith(selectedPage: pageIndex));
     log('Page switched to $pageIndex');
   }
